+++ CHANGELOG.md
@@ -1,3 +1,1 @@
-- created dimensional-weather project
-- added mixin code from [fabric-backrooms](https://github.com/crafter312/fabric-backrooms/) project
-- Quilt support requires use of Quilted Fabric API
+- refactored project to support dual fabric/forge versions